api/requirements.txt
fastapi[all]~=0.100
Pillow~=10.0
numpy~=1.26
requests~=2.31
tritonclient[http]~=2.39
opencv-python-headless~=4.7